Replace 'POST_SLUG' with the actual post slug, so posts with associated assets can link to them with generic Markdown.
This is intended to let the trailing-slash-free stuff work more smoothly, so I don't have to keep the slug up to date in the post text itself, and can just do [normal links like this](POST_SLUG/someAsset.png). It's kinda hacky, but it does the job. It'll probably break things when trailing slashes are left on, but really there's no sense using trailing slashes together with POST_SLUG, so I'm not going to worry about it. Caveat emptor.